added some exercises

+++ Java/officialPractices/factorial.java
@@ -6,9 +6,10 @@ public class factorial {
     public static void  main( String [] args){
       System.out.println("enter a number:");
       Scanner read=new Scanner(System.in);
-      int firstnum=read.nextInt();
-      int lastnum=1;
-      for(int i=firstnum; i>=1; i--){
+      long firstnum=read.nextInt();
+      long lastnum=1;
+      for(long i=firstnum; i>=1; i--){
+        // System.out.println(i);
             lastnum *= i;
           }
           read.close();
